+++ sphere-models/src/main/java/io/sphere/sdk/cartdiscounts/DiscountedLineItemPriceForQuantity.java
@@ -4,16 +4,16 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import io.sphere.sdk.models.Base;
 
 public class DiscountedLineItemPriceForQuantity extends Base {
-    private final Integer quantity;
+    private final Long quantity;
     private final DiscountedLineItemPrice discountedPrice;
 
     @JsonCreator
-    private DiscountedLineItemPriceForQuantity(final DiscountedLineItemPrice discountedPrice, final Integer quantity) {
+    private DiscountedLineItemPriceForQuantity(final DiscountedLineItemPrice discountedPrice, final Long quantity) {
         this.discountedPrice = discountedPrice;
         this.quantity = quantity;
     }
 
-    public static DiscountedLineItemPriceForQuantity of(final DiscountedLineItemPrice discountedLineItemPrice, final Integer quantity) {
+    public static DiscountedLineItemPriceForQuantity of(final DiscountedLineItemPrice discountedLineItemPrice, final Long quantity) {
         return new DiscountedLineItemPriceForQuantity(discountedLineItemPrice, quantity);
     }
 
@@ -21,7 +21,7 @@ public class DiscountedLineItemPriceForQuantity extends Base {
         return discountedPrice;
     }
 
-    public Integer getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 }
